Move NF4Tensor to quantization.quantize_.workflows
**Summary:**

Move `nf4tensor.py` to `torchao/quantization/quantize_/workflows/nf4/nf4_tensor.py`,
consistent with Int4Tensor, Int8Tensor, Float8Tensor, and other tensor types.
Move corresponding tests from `test/dtypes/` to
`test/quantization/quantize_/workflows/nf4/`.

Path migrations:
- torchao/dtypes/nf4tensor.py → torchao/quantization/quantize_/workflows/nf4/nf4_tensor.py
- test/dtypes/test_nf4.py → test/quantization/quantize_/workflows/nf4/test_nf4_tensor.py
- test/dtypes/ddp/ → test/quantization/quantize_/workflows/nf4/ddp/

After this commit, `torchao/dtypes/` is now empty — its `__init__.py` contains
only a stub `AffineQuantizedTensor` class for backward compatibility with
external libraries (diffusers, transformers). It will remain this way until
those libraries remove their AQT references, at which point `torchao/dtypes/`
can be deleted entirely.

New canonical imports:
- `from torchao.quantization import NF4Tensor, to_nf4`

NF4WeightOnlyConfig remains in `torchao/prototype/_nf4tensor_api.py` (out of
scope for this commit).

**Test Plan:**

```
pytest test/quantization/quantize_/workflows/nf4/test_nf4_tensor.py -x -v
```
